// src/app/models/members/legal.ts
export interface Legal{
    employeeCodeOfConductConfirmed: boolean;
    clause8aConfirmed: boolean;
    expandedCriminalRecordConfirmed: boolean;
    expandedCriminalRecordConfirmedDate: Date;
    expandedCriminalRecord: string; // The file it self
    communityMember: boolean;
    leaderSignUp: boolean;
    communityName: string;
    firstResponder: boolean;
    chainsawCourse: boolean;
    clause43IfSG: boolean;
    memberCanSwim: boolean;
    memberRightsOnOwnPicture: boolean;
    memberCanGoHome: boolean;
}